1.1

+++ PhoneShop/Controllers/UserController.cs
@@ -2,8 +2,9 @@
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
-using PhoneShop.Model;
-using PhoneShop.Services;
+using PhoneShop.Model.APIResponse;
+using PhoneShop.Model.User;
+using PhoneShop.Services.UserService;
 using System.Net;
 
 namespace PhoneShop.Controllers
@@ -16,6 +17,7 @@ namespace PhoneShop.Controllers
         private readonly IMapper _mapper;
         private readonly APIResponse _apiResponse;
         private readonly IUserService _userService;
+        // Constructor: injects dependencies for logging, mapping, and user service
         public UserController(ILogger<UserController> logger, IMapper mapper, IUserService userService)
         {
             _logger = logger;
@@ -23,6 +25,7 @@ namespace PhoneShop.Controllers
             _apiResponse = new APIResponse();
             _userService = userService;
         }
+        // create 1 user
         [HttpPost]
         [Route("CreateUser")]
         [ProducesResponseType(StatusCodes.Status200OK)]
@@ -32,23 +35,14 @@ namespace PhoneShop.Controllers
         [ProducesResponseType(StatusCodes.Status403Forbidden)]
         public async Task<ActionResult<APIResponse>> CreateUserAsync(RegisterDTO dto)
         {
-            try
-            {
-                var userCreated = await _userService.CreateUser(dto);
+            var userCreated = await _userService.CreateUser(dto);
 
-                _apiResponse.Data = userCreated;
-                _apiResponse.Status = true;
-                _apiResponse.StatusCode = HttpStatusCode.OK;
-                return Ok(_apiResponse);
-            }
-            catch (Exception ex)
-            {
-                _apiResponse.Status = false;
-                _apiResponse.StatusCode = HttpStatusCode.InternalServerError;
-                _apiResponse.Errors.Add(ex.Message);
-                return _apiResponse;
-            }
+            _apiResponse.Data = userCreated;
+            _apiResponse.Status = true;
+            _apiResponse.StatusCode = HttpStatusCode.OK;
+            return Ok(_apiResponse);
         }
+        // get all users
         [HttpGet]
         [Route("All", Name = "GetAllUser")]
         [ProducesResponseType(StatusCodes.Status200OK)]
@@ -58,25 +52,14 @@ namespace PhoneShop.Controllers
         [ProducesResponseType(StatusCodes.Status404NotFound)]
         public async Task<ActionResult<APIResponse>> GetUser()
         {
-            try
-            {
-                var users = await _userService.GetAllUsers();
+            var users = await _userService.GetAllUsers();
 
-                _apiResponse.Data = users;
-                _apiResponse.Status = true;
-                _apiResponse.StatusCode = HttpStatusCode.OK;
-                return Ok(_apiResponse);
-            }
-            catch (Exception ex)
-            {
-                _apiResponse.Status = false;
-                _apiResponse.StatusCode = HttpStatusCode.InternalServerError;
-                _apiResponse.Errors.Add(ex.Message);
-                return _apiResponse;
-            }
-
+            _apiResponse.Data = users;
+            _apiResponse.Status = true;
+            _apiResponse.StatusCode = HttpStatusCode.OK;
+            return Ok(_apiResponse);
         }
-
+        // get user by id
         [HttpGet]
         [Route("{id}", Name = "GetUsertById")]
         [ProducesResponseType(StatusCodes.Status200OK)]
@@ -87,38 +70,28 @@ namespace PhoneShop.Controllers
         [ProducesResponseType(StatusCodes.Status403Forbidden)]
         public async Task<ActionResult<APIResponse>> GetUserById(int id)
         {
-            try
+            if (id <= 0)
             {
-                if (id <= 0)
-                {
-                    _logger.LogWarning("Bad request for user id={Id}", id);
-                    return BadRequest();
-                }
-
-                var user = await _userService.GetUserById(id);
-                // NotFound - 404 - client error
-                if (user == null)
-                {
-                    _logger.LogError("user with id={Id} not found", id);
-                    return NotFound($"cant find user have id={id}");
-                }
-
-                _apiResponse.Data = user;
-                _apiResponse.Status = true;
-                _apiResponse.StatusCode = HttpStatusCode.OK;
-                // Ok - 200 - success
-                return Ok(_apiResponse);
+                _logger.LogWarning("Bad request for user id={Id}", id);
+                return BadRequest();
             }
-            catch (Exception ex)
+
+            var user = await _userService.GetUserById(id);
+            // NotFound - 404 - client error
+            if (user == null)
             {
-                _apiResponse.Status = false;
-                _apiResponse.StatusCode = HttpStatusCode.InternalServerError;
-                _apiResponse.Errors.Add(ex.Message);
-                return _apiResponse;
+                _logger.LogError("user with id={Id} not found", id);
+                return NotFound($"cant find user have id={id}");
             }
-            // BadRequest - 400 - client error
+
+            _apiResponse.Data = user;
+            _apiResponse.Status = true;
+            _apiResponse.StatusCode = HttpStatusCode.OK;
+            // Ok - 200 - success
+            return Ok(_apiResponse);
 
         }
+        // update user
         [HttpPut]
         [Route("Update")]
         [ProducesResponseType(StatusCodes.Status204NoContent)]
@@ -129,32 +102,24 @@ namespace PhoneShop.Controllers
         [Authorize]
         public async Task<ActionResult<APIResponse>> UpdateUser([FromBody] UserUpdateDTO dto)
         {
-            try
-            {
-                // role ?
-                var userIdClaim = User.FindFirst("UserId")?.Value;
-                var role = User.FindFirst(System.Security.Claims.ClaimTypes.Role)?.Value;
-                if(role != "Admin" && userIdClaim != dto.Id.ToString())
-                    return Forbid();
+            // check role
+            var userIdClaim = User.FindFirst("UserId")?.Value;
+            var role = User.FindFirst(System.Security.Claims.ClaimTypes.Role)?.Value;
+            if (role != "Admin" && userIdClaim != dto.Id.ToString())
+                return Forbid();
 
-                if (dto == null || dto.Id <= 0)
-                    return BadRequest();
-                var result = await _userService.UpdateUser(dto);
-                _apiResponse.Data = result;
-                _apiResponse.Status = true;
-                _apiResponse.StatusCode = HttpStatusCode.OK;
-                return Ok(_apiResponse);
-            }
-            catch (Exception ex)
-            {
-                _apiResponse.Status = false;
-                _apiResponse.StatusCode = HttpStatusCode.InternalServerError;
-                _apiResponse.Errors ??= new List<string>();
-                _apiResponse.Errors.Add(ex.Message);
-                return _apiResponse;
-            }
+            // validate dto
+            if (dto == null || dto.Id <= 0)
+                return BadRequest();
+            // call service to update user
+            var result = await _userService.UpdateUser(dto);
+            _apiResponse.Data = result;
+            _apiResponse.Status = true;
+            _apiResponse.StatusCode = HttpStatusCode.OK;
+            return Ok(_apiResponse);
 
         }
+        // delete user by id
         [HttpDelete("Delete/{id}", Name = "DeleteUserById")]
         [ProducesResponseType(StatusCodes.Status200OK)]
         [ProducesResponseType(StatusCodes.Status400BadRequest)]
@@ -165,23 +130,20 @@ namespace PhoneShop.Controllers
         [Authorize(Roles = "Admin")]
         public async Task<ActionResult<APIResponse>> DeleteUser(int id)
         {
-            try
+            // check if id is valid
+            var user = await _userService.GetUserById(id);
+            // NotFound - 404 - client error
+            if (user == null)
             {
-                var result = await _userService.DeleteUser(id);
-
-                _apiResponse.Data = result;
-                _apiResponse.Status = true;
-                _apiResponse.StatusCode = HttpStatusCode.OK;
-                return Ok(_apiResponse);
+                _logger.LogError("user with id={Id} not found", id);
+                return NotFound($"cant find user have id={id}");
             }
-            catch (Exception ex)
-            {
-                _apiResponse.Status = false;
-                _apiResponse.StatusCode = HttpStatusCode.InternalServerError;
-                _apiResponse.Errors ??= new List<string>();
-                _apiResponse.Errors.Add(ex.Message);
-                return _apiResponse;
-            }
+            // call service to delete user
+            var result = await _userService.DeleteUser(id);
+            _apiResponse.Data = result;
+            _apiResponse.Status = true;
+            _apiResponse.StatusCode = HttpStatusCode.OK;
+            return Ok(_apiResponse);
 
         }
     }
